Move OptionList and OptionButton to utils
